+++ changelog.md
@@ -37,8 +37,13 @@
  - video: archiving
  - video: previews
 
+**11/10/2018**
+ - socket.io infrastructure
+ 
 **1.1.0**
  - next live @ & schedule
  - chat log
+
+
 **2.0.0 : Android App : ??/??/??**
  - android rtmp relay?